+++ lib/sbmt/strangler/action.rb
@@ -8,7 +8,7 @@ module Sbmt
       option :params_tracking_allowlist, :headers_allowlist, :flipper_actor, default_from: :controller
       option :proxy_url, :proxy_http_verb
       option :mirror, default: ->(_rails_controller) {}
-      option :mirror_compare, default: ->(_origin_response, _mirror_result) { false }
+      option :compare, default: ->(_origin_response, _mirror_result) { false }
 
       attr_reader :name, :controller