+++ lib/sbmt/strangler/action_invoker.rb
@@ -3,22 +3,25 @@
 module Sbmt
   module Strangler
     class ActionInvoker
-      attr_reader :rails_controller, :strangler_action, :feature_flags
+      attr_reader :rails_controller, :strangler_action, :metric_tracker, :feature_flags
 
-      def initialize(rails_controller:, strangler_action:, feature_flags:)
+      def initialize(rails_controller:, strangler_action:, metric_tracker:, feature_flags:)
         @rails_controller = rails_controller
         @strangler_action = strangler_action
+        @metric_tracker = metric_tracker
         @feature_flags = feature_flags
       end
 
-      delegate :track_params_usage, :track_work_mode, to: :rails_controller
+      delegate :track_params_usage, :track_work_mode, :log_unallowed_params, to: :metric_tracker
 
       def call
         track_params_usage
+        log_unallowed_params
         track_work_mode(work_mode_class.name.demodulize.underscore)
         work_mode_class.new(
           rails_controller:,
           strangler_action:,
+          metric_tracker:,
           feature_flags:
         ).call
       end
@@ -27,9 +30,10 @@ module Sbmt
 
       def work_mode_class
         @work_mode_class ||=
-          if feature_flags.replace?
-            Sbmt::Strangler::WorkModes::Replace
-          elsif feature_flags.mirror?
+          # TODO: enabled 'replace' work mode
+          # if feature_flags.replace?
+          #   Sbmt::Strangler::WorkModes::Replace
+          if feature_flags.mirror?
             Sbmt::Strangler::WorkModes::Mirror
           else
             Sbmt::Strangler::WorkModes::Proxy