+++ lib/sbmt/strangler/feature_flags.rb
@@ -5,8 +5,7 @@ module Sbmt
     class FeatureFlags
       FLAGS = %i[
         mirror
-        replace
-      ]
+      ] # TODO: add 'replace' feature flag
 
       attr_reader :strangler_action, :rails_controller