+++ spec/internal/config/initializers/strangler.rb
@@ -12,7 +12,7 @@ Sbmt::Strangler.configure do |strangler|
       action.proxy_url = "http://example.com:8080/api/stores"
       action.proxy_http_verb = :post
       action.mirror = ->(_rails_controller) { '["mirror_result"]' }
-      action.mirror_compare = ->(origin_response, mirror_result) {
+      action.compare = ->(origin_response, mirror_result) {
         mirror_result == '["mirror_result"]' && origin_response == '["origin_response"]'
       }
     end